refactor(memory_consolidation): promote to directory, split tests (#1266)
- src/memory_consolidation.rs: 525 LOC → src/memory_consolidation/mod.rs (323 LOC)
- New src/memory_consolidation/tests.rs (202 LOC)

Pure refactor: 11 tests pass.
